+++ rossum-agent/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ---
 
-## [2.0.0] - 2026-04-08
+## [2.0.0] - 2026-04-09
 
 ### Added
 - Allow `http://localhost:3000` CORS origin when `ROSSUM_AGENT_DEVELOP` env var is set